+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Upcoming
 
 * Scale: `Local origin` is now only shown if applying on a layer. The image resolution values are only shown if `Set image resolution in DPI` is checked.
+* Rotate: `Interpolation` is now only shown for a custom angle as it is otherwise unused.
 * Moved less common options (`Filter`, `Apply non-destructively`, `Blend mode`) for Brightness-Contrast, Levels and Curves under More options.